chore: consolidate content files
