+++ python/mirascope/llm/content/content.py
@@ -1,10 +1,11 @@
 from typing import TypeAlias
 
-from .audio import Audio, AudioUrl
+from .audio import Audio, AudioChunk, AudioUrl
+from .chunk_boundary import ChunkEnd, ChunkStart
 from .document import Document
-from .image import Image, ImageUrl
-from .thinking import Thinking
-from .tool_call import ToolCall
+from .image import Image, ImageChunk, ImageUrl
+from .thinking import Thinking, ThinkingChunk
+from .tool_call import ToolCall, ToolCallChunk
 from .tool_output import ToolOutput
 
 UserContent: TypeAlias = (
@@ -17,3 +18,14 @@ AssistantContent: TypeAlias = str | Image | Audio | ToolCall | Thinking
 
 SystemContent: TypeAlias = str
 """Content types that can be included in a SystemMessage."""
+
+ContentChunk: TypeAlias = (
+    str
+    | ImageChunk
+    | AudioChunk
+    | ToolCallChunk
+    | ThinkingChunk
+    | ChunkStart
+    | ChunkEnd
+)
+"""Content chunk types that may be emitted by Streams."""
